separated routes for mugs and mug_pics, added todo for get at id
